fastapiapp/requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
boto3==1.26.76
botocore==1.29.76
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.0
fastapi==0.92.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
numpy==1.24.2
orjson==3.8.6
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
sqlite-s3-query==0.0.74
starlette==0.25.0
tomli==2.0.1
typing-extensions==4.5.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4